Factorized document processing code
