+++ build_tools/requirements_pip_win.txt
@@ -6,5 +6,6 @@ flask_apispec
 # disabling it, version 0.6 currently broken for sparse arrays https://github.com/dmlc/xgboost/issues/1238#issuecomment-247787192 
 nltk
 scikit-learn==0.18.1  
+git+https://github.com/FreeDiscovery/jwzthreading.git
 # gensim # not used at present
 # coverage # not used at present
